renamed module for storage classes for comprehensibility
